ion: system_heap: fix a constant-conversion warning
../drivers/staging/android/ion/ion_system_secure_heap.c:417:22: warning: implicit conversion from 'unsigned long' to 'int' changes value from 18446744073709551615 to -1 [-Wconstant-conversion]
        for (i = NUM_ORDERS - 2; i >= 0; i--) {
               ~ ~~~~~~~~~~~^~~
1 warning generated.

+++ drivers/staging/android/ion/ion_system_heap.h
@@ -25,7 +25,7 @@ static const unsigned int orders[] = {9, 4, 0};
 static const unsigned int orders[] = {0};
 #endif
 
-#define NUM_ORDERS ARRAY_SIZE(orders)
+#define NUM_ORDERS (int)ARRAY_SIZE(orders)
 
 struct ion_system_heap {
 	struct ion_heap heap;
